More moves
Extend solving range
